refactor(relay): rename pkg name to autorelay
