+++ internal/pkg/conn/p2p/relayfilter.go
@@ -3,7 +3,7 @@ package p2p
 import (
 	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/rumsystem/quorum/internal/pkg/storage"
-	"github.com/rumsystem/quorum/pkg/relayapi/handlers"
+	"github.com/rumsystem/quorum/pkg/autorelay/handlers"
 
 	ma "github.com/multiformats/go-multiaddr"
 )
